Doclynx/DocLynx/requirements.txt
flask
faiss-cpu
sentence-transformers
pandas